+++ src/app/app-routing.module.ts
@@ -6,6 +6,7 @@ import { TasksComponent } from './pages/tasks/tasks.component';
 import { ClientsComponent } from './pages/clients/clients.component';
 import { ProfileComponent } from './pages/profile/profile.component';
 import { IsLoggedGuard } from './guards/is-logged.guard';
+import { EditprofileComponent } from './pages/profile/editprofile/editprofile.component';
 
 const routes: Routes = [
   { path: '', component: DashboardComponent, data: {title: 'Dashboard'}},
@@ -13,6 +14,8 @@ const routes: Routes = [
   { path: 'projects', component: ProjectsComponent, data: {title: 'Projects'}, canActivate: [IsLoggedGuard]},
   { path: 'tasks', component: TasksComponent, data: {title: 'Tasks'}, canActivate: [IsLoggedGuard]},
   { path: 'profile', component: ProfileComponent, data: {title: 'Profile'}, canActivate: [IsLoggedGuard]},
+  { path: 'edit-profile', component: EditprofileComponent, data: {title: 'Edit Profile'}, canActivate: [IsLoggedGuard]},
+
 
 ];